+++ Runtime/Scripts/Setup/Settings.cs
@@ -53,7 +53,7 @@ namespace GameAnalyticsSDK.Setup
         /// The version of the GA Unity Wrapper plugin
         /// </summary>
         [HideInInspector]
-        public static string VERSION = "7.10.6";
+        public static string VERSION = "8.0.0";
 
         [HideInInspector]
         public static bool CheckingForUpdates = false;