+++ Runtime/Scripts/Wrapper/GA_MonoWrapper.cs
@@ -7,7 +7,7 @@ namespace GameAnalyticsSDK.Wrapper
 {
     public partial class GA_Wrapper
     {
-#if (UNITY_STANDALONE || UNITY_WP_8_1 || UNITY_SAMSUNGTV) && (!UNITY_EDITOR)
+#if GA_USE_MONO_WRAPPER && ((UNITY_STANDALONE || UNITY_WP_8_1 || UNITY_SAMSUNGTV) && (!UNITY_EDITOR))
 
         private class UnityRemoteConfigsListener : GameAnalyticsSDK.Net.IRemoteConfigsListener
         {