+++ Runtime/Scripts/Wrapper/GA_UWPWrapper.cs
@@ -7,6 +7,7 @@ using System.Runtime.InteropServices;
 
 namespace GameAnalyticsSDK.Wrapper
 {
+    /*
     public partial class GA_Wrapper
     {
 
@@ -272,4 +273,5 @@ namespace GameAnalyticsSDK.Wrapper
         }
 #endif
     }
+    */
 }